src/moca_modules/requirements.txt
pytz
GitPython
aiosmtplib
PyMySQL
aiomysql
aiofiles
aioredis
requests
pymysql-pooling
sanic
ujson
aiohttp
cchardet
aiodns
docopt
pycryptodome
leveldb
twilio
limits
redis
mycli
litecli
setproctitle